+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [CLI](projects%2Fcli%2FCHANGELOG.md)
 - [Design System](projects%2Fdesign-system%2FCHANGELOG.md)
 - [Oblique](projects%2Foblique%2FCHANGELOG.md)
 - [Sandbox](projects%2Fsandbox%2FCHANGELOG.md)